# Minor Optimization to Occlusion Culling