Minor performance improvement to occlusion culling as well as a partial fix for #106184 (I'm working on something more comprehensive). 

The performance gain comes from swapping one version of `Projection::xform` for another that is inlined. The two functions aren't a 1-to-1 match, so a minor rework the logic was required. After further testing, I believe the issue of small objects self-occluding (as described by @JFonS in #52545) was fully resolved in #94210. I’ve therefore updated the logic to be less conservative, resulting in a small improvement in occlusion rate which can be seen in the video below. The test [project](https://github.com/user-attachments/files/20847878/occlusion-base.zip) is similar to the MRP in #106184. The camera is placed inside a box occluder, so ideally, everything should become occluded.

https://github.com/user-attachments/assets/30db2056-3115-4348-8b88-3556c7020dfa

Regarding the performance improvements, I tested using [occlusion_culling_mesh_lod](https://github.com/godotengine/godot-demo-projects/tree/master/3d/occlusion_culling_mesh_lod). On my machine, it went from roughly ~1.2 ms/frame (~830 FPS) to ~1.05 ms/frame (~950 FPS). Since this project is specifically designed to be bottlenecked by occlusion culling, I also tested using the [tps-demo](https://github.com/godotengine/tps-demo#). However, I wasn’t able to observe any noticeable performance improvement (~122 FPS before and after).

>Godot v4.5.beta (b3d858fad) - Ubuntu 24.04.2 LTS 24.04 on X11 - X11 display driver, Multi-window, 2 monitors - Vulkan (Forward+) - dedicated NVIDIA GeForce RTX 2070 SUPER (nvidia; 550.144.03) - AMD Ryzen 7 3700X 8-Core Processor (16 threads) - 14.76 GiB memory